CRUD em tarefa

+++ GerenciadorTarefas/GerenciadorTarefas/Domain/Repository/TarefaRepository.cs
@@ -8,39 +8,38 @@ namespace GerenciadorTarefas.Domain.Repository
     public class TarefaRepository : ITarefaRepository
     {
         private readonly Context _context;
-        public TarefaRepository(Context context) { _context = context; }
+
+        public TarefaRepository(Context context)
+        {
+            _context = context;
+        }
 
         public async Task<Tarefa> Create(Tarefa tarefa)
         {
-            var ret = await _context.Tarefas.AddAsync(tarefa);
-
+            var result = await _context.Tarefas.AddAsync(tarefa);
             await _context.SaveChangesAsync();
-
-            ret.State = EntityState.Detached;
-
-            return ret.Entity;
+            return result.Entity;
         }
 
         public async Task<bool> Delete(int tarefaId)
         {
             var tarefa = await _context.Tarefas.FindAsync(tarefaId);
+            if (tarefa == null)
+                return false;
+
             _context.Tarefas.Remove(tarefa);
-
             await _context.SaveChangesAsync();
-
             return true;
         }
 
         public async Task<List<Tarefa>> List()
         {
-            List<Tarefa> list = await _context.Tarefas.OrderBy(p => p.Titulo).ToListAsync();
-            return list;
+            return await _context.Tarefas.ToListAsync();
         }
 
         public async Task<int> Update(Tarefa tarefa)
         {
             _context.Entry(tarefa).State = EntityState.Modified;
-
             return await _context.SaveChangesAsync();
         }
     }
